3.2.0

+++ package/kedro_viz/server.py
@@ -1,4 +1,4 @@
-# Copyright 2018-2019 QuantumBlack Visual Analytics Limited
+# Copyright 2020 QuantumBlack Visual Analytics Limited
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -31,17 +31,21 @@
 import hashlib
 import json
 import multiprocessing
-import sys
+import socket
 import webbrowser
 from collections import defaultdict
+from contextlib import closing
 from pathlib import Path
 from typing import Dict
 
 import click
+import kedro
 import requests
 from flask import Flask, jsonify, send_from_directory
 from IPython.core.display import HTML, display
-from kedro.cli import get_project_context
+from kedro.cli import get_project_context  # pylint: disable=ungrouped-imports
+from kedro.cli.utils import KedroCliError  # pylint: disable=ungrouped-imports
+from semver import match
 
 from kedro_viz.utils import wait_for
 
@@ -51,6 +55,17 @@ data = None  # pylint: disable=invalid-name
 
 app = Flask(  # pylint: disable=invalid-name
     __name__, static_folder=str(Path(__file__).parent.absolute() / "html" / "static")
+)
+
+ERROR_PROJECT_ROOT = (
+    "Could not find a Kedro project root. You can run `kedro viz` by either providing "
+    "`--load-file` flag with a filepath to a JSON pipeline representation, "
+    "or if the current working directory is the root of a Kedro project."
+)
+
+ERROR_PIPELINE_FLAG_NOT_SUPPORTED = (
+    "`--pipeline` flag was provided, but it is not supported "
+    "in Kedro version {}".format(kedro.__version__)
 )
 
 
@@ -69,7 +84,11 @@ def _hash(value):
 
 def _check_viz_up(port):
     url = "http://127.0.0.1:{}/".format(port)
-    response = requests.get(url)
+    try:
+        response = requests.get(url)
+    except requests.ConnectionError:
+        return False
+
     return response.status_code == 200
 
 
@@ -84,10 +103,10 @@ def run_viz(port=None, line=None) -> None:
         line: line required by line magic interface.
 
     """
-    if not port:  # Default argument doesn't work in Jupyter line magic
-        port = 4141
+    port = port or 4141  # Default argument doesn't work in Jupyter line magic
+    port = _allocate_port(start_at=port)
 
-    if port in _VIZ_PROCESSES:
+    if port in _VIZ_PROCESSES and _VIZ_PROCESSES[port].is_alive():
         _VIZ_PROCESSES[port].terminate()
 
     viz_process = multiprocessing.Process(
@@ -105,6 +124,59 @@ def run_viz(port=None, line=None) -> None:
         port
     )
     display(HTML(wrapper))
+
+
+def _allocate_port(start_at: int, end_at: int = 65535) -> int:
+    acceptable_ports = range(start_at, end_at + 1)
+
+    viz_ports = _VIZ_PROCESSES.keys() & set(acceptable_ports)
+    if viz_ports:  # reuse one of already allocated ports
+        return sorted(viz_ports)[0]
+
+    socket.setdefaulttimeout(2.0)  # seconds
+    for port in acceptable_ports:  # iterate through all acceptable ports
+        with closing(socket.socket(socket.AF_INET, socket.SOCK_STREAM)) as sock:
+            if sock.connect_ex(("127.0.0.1", port)) != 0:  # port is available
+                return port
+
+    raise ValueError(
+        "Cannot allocate an open TCP port for Kedro-Viz in a range "
+        "from {} to {}".format(start_at, end_at)
+    )
+
+
+def _load_from_file(load_file: str) -> dict:
+    global data  # pylint: disable=global-statement,invalid-name
+    data = json.loads(Path(load_file).read_text())
+    for key in ["nodes", "edges", "tags"]:
+        if key not in data:
+            raise KedroCliError(
+                "Invalid file, top level key '{}' not found.".format(key)
+            )
+    return data
+
+
+def _get_pipeline_from_context(context, pipeline_name):
+    if match(kedro.__version__, ">=0.15.2"):
+        return context._get_pipeline(  # pylint: disable=protected-access
+            name=pipeline_name
+        )
+    # Kedro 0.15.0 or 0.15.1
+    if pipeline_name:
+        raise KedroCliError(ERROR_PIPELINE_FLAG_NOT_SUPPORTED)
+    return context.pipeline
+
+
+def _get_pipeline_catalog_from_kedro14(env):
+    try:
+        pipeline = get_project_context("create_pipeline")()
+        get_config = get_project_context("get_config")
+        conf = get_config(str(Path.cwd()), env)
+        create_catalog = get_project_context("create_catalog")
+        catalog = create_catalog(config=conf)
+        return pipeline, catalog
+    except (ImportError, KeyError):
+        raise KedroCliError(ERROR_PROJECT_ROOT)
 
 
 def format_pipeline_data(pipeline, catalog):
@@ -177,6 +249,7 @@ def commands():
     """Visualize the pipeline using kedroviz."""
 
 
+# pylint: disable=too-many-arguments
 @commands.command(context_settings=dict(help_option_names=["-h", "--help"]))
 @click.option(
     "--host",
@@ -195,27 +268,76 @@ def commands():
     help="Whether to open viz interface in the default browser or not. "
     "Defaults to True.",
 )
-@click.option("--load-file", default=None, type=click.Path(exists=True, dir_okay=False))
 @click.option(
-    "--save-file", default=None, type=click.Path(dir_okay=False, writable=True)
+    "--load-file",
+    default=None,
+    type=click.Path(exists=True, dir_okay=False),
+    help="Path to load the pipeline JSON file",
 )
-def viz(host, port, browser, load_file, save_file):
+@click.option(
+    "--save-file",
+    default=None,
+    type=click.Path(dir_okay=False, writable=True),
+    help="Path to save the pipeline JSON file",
+)
+@click.option(
+    "--pipeline",
+    type=str,
+    default=None,
+    help="Name of the modular pipeline to visualize. "
+    "If not set, the default pipeline is visualized",
+)
+@click.option(
+    "--env",
+    "-e",
+    type=str,
+    default=None,
+    multiple=False,
+    envvar="KEDRO_ENV",
+    help="Kedro configuration environment. If not specified, "
+    "catalog config in `local` will be used",
+)
+def viz(host, port, browser, load_file, save_file, pipeline, env):
     """Visualize the pipeline using kedroviz."""
-    _call_viz(host, port, browser, load_file, save_file)
+    try:
+        _call_viz(host, port, browser, load_file, save_file, pipeline, env)
+    except KedroCliError:
+        raise
+    except Exception as ex:
+        raise KedroCliError(str(ex))
 
 
-def _call_viz(host=None, port=None, browser=None, load_file=None, save_file=None):
+# pylint: disable=too-many-arguments
+def _call_viz(
+    host=None,
+    port=None,
+    browser=None,
+    load_file=None,
+    save_file=None,
+    pipeline_name=None,
+    env=None,
+):
     global data  # pylint: disable=global-statement,invalid-name
 
     if load_file:
-        data = json.loads(Path(load_file).read_text())
-        for key in ["nodes", "edges", "tags"]:
-            if key not in data:
-                click.echo("Invalid file, top level key '{}' not found.".format(key))
-                sys.exit(1)
+        data = _load_from_file(load_file)
     else:
-        pipeline = get_project_context("create_pipeline")()
-        catalog = get_project_context("create_catalog")(None)
+        if match(kedro.__version__, ">=0.15.0"):
+            from kedro.context import KedroContextError
+
+            try:
+                context = get_project_context("context", env=env)
+                pipeline = _get_pipeline_from_context(context, pipeline_name)
+            except KedroContextError:
+                raise KedroCliError(ERROR_PROJECT_ROOT)
+            catalog = context.catalog
+
+        else:
+            # Kedro 0.14.*
+            if pipeline_name:
+                raise KedroCliError(ERROR_PIPELINE_FLAG_NOT_SUPPORTED)
+            pipeline, catalog = _get_pipeline_catalog_from_kedro14(env)
+
         data = format_pipeline_data(pipeline, catalog)
 
     if save_file:
